+++ docs/news.md
@@ -1,5 +1,9 @@
 # Changelog
 
+# coronavirus 0.0.6
+
+- Add total numbers as returned by Weixin. :warning: great discrepancy between those and the daily figures.
+
 ## coronavirus 0.0.5
 
 - Add hubei default province in DXY tab.